refactor(game-result): 플레이어 승패 계산 메소드를 GameResult로 이동
